# swagger-spec.json
{"openapi":"3.0.0","info":{"title":"Aso Books Api","description":"Api for Aso Books","version":"1.0","contact":{}},"tags":[{"name":"Asobooks","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"associationName":{"type":"string"},"password":{"type":"string","minLength":6}},"required":["firstName","lastName","email","associationName","password"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"ChangePasswordDto":{"type":"object","properties":{"password":{"type":"string"}},"required":["password"]},"PasswordResetDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ProfileUpdateDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"},"newPassword":{"type":"string"}}},"AssociationAddressRequestDto":{"type":"object","properties":{"address":{"type":"string"},"countryCode":{"type":"string"},"unit":{"type":"string"}},"required":["address","countryCode"]},"MemberSignUpDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"address":{"$ref":"#/components/schemas/AssociationAddressRequestDto"},"types":{"type":"array","items":{"type":"string","enum":["EXECUTIVE_ACCOUNT","MEMBER_ACCOUNT"]}},"identifier":{"type":"string"}},"required":["firstName","lastName","email","phoneNumber","types"]},"AddressUpdateDto":{"type":"object","properties":{"address":{"type":"string"},"unit":{"type":"string"},"countryCode":{"type":"string"}},"required":["address","unit","countryCode"]},"EditMemberDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressUpdateDto"},"phoneNumber":{"type":"string"},"types":{"type":"array","items":{"type":"string","enum":["EXECUTIVE_ACCOUNT","MEMBER_ACCOUNT"]}}},"required":["firstName","lastName","address","phoneNumber","types"]},"BankInfoRequestDto":{"type":"object","properties":{"code":{"type":"string"},"accountNumber":{"type":"string"}},"required":["code","accountNumber"]},"AssociationOnboardingDto":{"type":"object","properties":{"file":{"type":"string","format":"binary"},"name":{"type":"string"},"type":{"type":"string","enum":["ESTATE","CLUB","COOPERATIVE","UNION","ALUMNI","HOUSING"]},"address":{"$ref":"#/components/schemas/AssociationAddressRequestDto"},"bankInfo":{"$ref":"#/components/schemas/BankInfoRequestDto"}},"required":["name","type"]},"AssociationAddressUpdateRequestDto":{"type":"object","properties":{"address":{"type":"string"},"countryCode":{"type":"string"}},"required":["countryCode"]},"UpdateAssociationDto":{"type":"object","properties":{"name":{"type":"string"},"address":{"$ref":"#/components/schemas/AssociationAddressUpdateRequestDto"},"type":{"type":"string","enum":["ESTATE","CLUB","COOPERATIVE","UNION","ALUMNI","HOUSING"]},"bankInfo":{"$ref":"#/components/schemas/BankInfoRequestDto"}}},"ServiceFeeRequestDto":{"type":"object","properties":{"name":{"type":"string"},"amountInMinorUnit":{"type":"number","minimum":0},"description":{"type":"string"},"type":{"type":"string","enum":["ONE_TIME","RE_OCCURRING"]},"cycle":{"type":"string","enum":["WEEKLY","BI_WEEKLY","MONTHLY","QUARTERLY","YEARLY","ONE_OFF"]},"billingStartDate":{"type":"string"},"dueDate":{"type":"string"},"recipients":{"type":"array","items":{"type":"string"}}},"required":["name","amountInMinorUnit","description","type"]},"MembershipFeeRequestDto":{"type":"object","properties":{"memberIdentifiers":{"type":"array","items":{"type":"string"}}},"required":["memberIdentifiers"]},"InvoiceRequestDto":{"type":"object","properties":{"billCodes":{"type":"array","items":{"type":"string"}}},"required":["billCodes"]},"RoleMembershipRequestDto":{"type":"object","properties":{"membershipReferences":{"type":"array","items":{"type":"string"}}},"required":["membershipReferences"]},"RoleRequest":{"type":"object","properties":{}},"WalletWithdrawalDto":{"type":"object","properties":{"amountInMinorUnit":{"type":"number","minimum":1},"password":{"type":"string"}},"required":["amountInMinorUnit","password"]}}},"paths":{"/api/v1/test":{"get":{"operationId":"TestController_index","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/test/association":{"get":{"operationId":"TestController_getAssociation","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/sign-up":{"post":{"operationId":"AuthenticationController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/validate-principal/{token}":{"get":{"operationId":"AuthenticationController_principalSetUp","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/login":{"post":{"operationId":"AuthenticationController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/password/reset/{token}":{"post":{"operationId":"AuthenticationController_onPasswordReset","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/password/reset":{"post":{"operationId":"AuthenticationController_passwordReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/me":{"get":{"operationId":"AuthenticationController_me","parameters":[],"responses":{"200":{"description":""}}},"patch":{"operationId":"AuthenticationController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateDto"}}}},"responses":{"200":{"description":""}}}},"/api/v1/validate-principal":{"post":{"operationId":"AuthenticationController_sendVerificationToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/membership-management/create":{"post":{"operationId":"MembershipManagementController_createAssociationMember","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberSignUpDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/membership-management/{identifier}":{"get":{"operationId":"MembershipManagementController_getMember","parameters":[{"name":"identifier","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"MembershipManagementController_deleteMember","parameters":[{"name":"identifier","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"MembershipManagementController_updateMember","parameters":[{"name":"identifier","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditMemberDto"}}}},"responses":{"200":{"description":""}}}},"/api/v1/membership-management":{"get":{"operationId":"MembershipManagementController_getAssociationMembers","parameters":[{"name":"query","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"enum":["EXECUTIVE_ACCOUNT","MEMBER_ACCOUNT"],"type":"string"}},{"name":"dateCreatedBefore","required":true,"in":"query","schema":{"type":"string"}},{"name":"dateCreatedAfter","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"offset","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/v1/associations/onboard":{"put":{"operationId":"AssociationController_createAssociation","parameters":[],"requestBody":{"required":true,"description":"Onboard associations","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/AssociationOnboardingDto"}}}},"responses":{"200":{"description":""}}}},"/api/v1/associations":{"get":{"operationId":"AssociationController_getAssociation","parameters":[],"responses":{"200":{"description":""}}},"patch":{"operationId":"AssociationController_update","parameters":[],"requestBody":{"required":true,"description":"Update Association","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UpdateAssociationDto"}}}},"responses":{"200":{"description":""}}}},"/api/v1/master-records/banks":{"get":{"operationId":"MasterRecordController_getBanks","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/master-records/counties":{"get":{"operationId":"MasterRecordController_getCountries","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/member-workspace/bills":{"get":{"operationId":"BillsController_getBills","parameters":[{"name":"minAmount","required":false,"in":"query","schema":{"type":"number"}},{"name":"maxAmount","required":false,"in":"query","schema":{"type":"number"}},{"name":"paymentStatus","required":false,"in":"query","schema":{"enum":["PAID","NOT_PAID","PARTIALLY_PAID","FAILED","PENDING","AWAITING_CONFIRMATION"],"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}},{"name":"serviceType","required":false,"in":"query","schema":{"enum":["ONE_TIME","RE_OCCURRING"],"type":"string"}},{"name":"startDateBefore","required":false,"in":"query","schema":{"type":"string"}},{"name":"dateStartAfter","required":false,"in":"query","schema":{"type":"string"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"phonenumber","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/service-fees":{"post":{"operationId":"ServiceFeeController_createService","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceFeeRequestDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"ServiceFeeController_all","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}},{"name":"frequency","required":false,"in":"query","schema":{"enum":["WEEKLY","BI_WEEKLY","MONTHLY","QUARTERLY","YEARLY","ONE_OFF"],"type":"string"}},{"name":"type","required":false,"in":"query","schema":{"enum":["ONE_TIME","RE_OCCURRING"],"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"enum":["ACTIVE","IN_ACTIVE","DELETED","PENDING_ACTIVATION","SUSPENDED"],"type":"string"}},{"name":"dateCreatedAfter","required":false,"in":"query","schema":{"type":"string"}},{"name":"dateCreatedBefore","required":false,"in":"query","schema":{"type":"string"}},{"name":"startDateBefore","required":false,"in":"query","schema":{"type":"string"}},{"name":"startDateAfter","required":false,"in":"query","schema":{"type":"string"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"minAmount","required":false,"in":"query","schema":{"type":"number"}},{"name":"maxAmount","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/v1/service-fees/{code}":{"delete":{"operationId":"ServiceFeeController_deActivateFee","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"get":{"operationId":"ServiceFeeController_getServiceByCode","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/service-fees/{code}/subscriptions":{"get":{"operationId":"ServiceFeeController_getSubscriptions","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"startDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}},{"name":"amountReceivedInMinorUnitGreater","required":false,"in":"query","schema":{"type":"number"}},{"name":"amountReceivedInMinorUnitLess","required":false,"in":"query","schema":{"type":"number"}},{"name":"amountPendingInMinorUnitLess","required":false,"in":"query","schema":{"type":"number"}},{"name":"amountPendingInMinorUnitGreater","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/v1/service-fees/{code}/bills":{"get":{"operationId":"ServiceFeeController_getBills","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"createdDateAfter","required":false,"in":"query","schema":{"type":"string"}},{"name":"createdDateBefore","required":false,"in":"query","schema":{"type":"string"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"phoneNumber","required":false,"in":"query","schema":{"type":"string"}},{"name":"timeOfPaymentBefore","required":false,"in":"query","schema":{"type":"string"}},{"name":"timeOfPaymentAfter","required":false,"in":"query","schema":{"type":"string"}},{"name":"receiptNumber","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}},{"name":"paymentStatus","required":false,"in":"query","schema":{"enum":["PAID","NOT_PAID","PARTIALLY_PAID","FAILED","PENDING","AWAITING_CONFIRMATION"],"type":"string"}},{"name":"feeType","required":false,"in":"query","schema":{"enum":["ONE_TIME","RE_OCCURRING"],"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/service-fees/{code}/members":{"patch":{"operationId":"GroupServiceFeeController_addMember","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MembershipFeeRequestDto"}}}},"responses":{"200":{"description":""}}},"get":{"operationId":"GroupServiceFeeController_getMembers","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"query","required":true,"in":"query","schema":{"type":"string"}},{"name":"type","required":true,"in":"query","schema":{"enum":["EXECUTIVE_ACCOUNT","MEMBER_ACCOUNT"],"type":"string"}},{"name":"dateCreatedBefore","required":true,"in":"query","schema":{"type":"string"}},{"name":"dateCreatedAfter","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"offset","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/v1/service-fees/{serviceCode}/members":{"delete":{"operationId":"GroupServiceFeeController_removeMember","parameters":[{"name":"serviceCode","required":true,"in":"path","schema":{"type":"string"}},{"name":"memberIdentifier","required":true,"in":"query","schema":{"type":"array"}}],"responses":{"200":{"description":""}}}},"/api/v1/verifications/banks/{code}/account/{accountNumber}/verify":{"get":{"operationId":"VerificationController_verifyBankAccountNumber","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"accountNumber","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/invoice/{code}/payment-request":{"get":{"operationId":"InvoiceController_getPaymentLink","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/invoice":{"post":{"operationId":"InvoiceController_createInvoice","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceRequestDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/payments/confirm":{"get":{"operationId":"PaymentRequestController_confirmPayment","parameters":[{"name":"reference","required":true,"in":"query","schema":{"type":"string"}},{"name":"merchantReference","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/dashboard/contribution-graph":{"get":{"operationId":"DashboardController_contributionGraph","parameters":[{"name":"year","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/v1/dashboard":{"get":{"operationId":"DashboardController_dashBoardStarts","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/activities":{"get":{"operationId":"ActivitiesController_recentActivities","parameters":[{"name":"type","required":true,"in":"query","schema":{"enum":["USER_MANAGEMENT","PAYMENT","AUTHENTICATION","SERVICE_FEE_AND_BILL_GENERATION"],"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/v1/subscriptions/{code}":{"get":{"operationId":"SubscriptionController_get","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"createdDateAfter","required":false,"in":"query","schema":{"type":"string"}},{"name":"createdDateBefore","required":false,"in":"query","schema":{"type":"string"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}},{"name":"phoneNumber","required":false,"in":"query","schema":{"type":"string"}},{"name":"timeOfPaymentBefore","required":false,"in":"query","schema":{"type":"string"}},{"name":"timeOfPaymentAfter","required":false,"in":"query","schema":{"type":"string"}},{"name":"receiptNumber","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}},{"name":"paymentStatus","required":false,"in":"query","schema":{"enum":["PAID","NOT_PAID","PARTIALLY_PAID","FAILED","PENDING","AWAITING_CONFIRMATION"],"type":"string"}},{"name":"feeType","required":false,"in":"query","schema":{"enum":["ONE_TIME","RE_OCCURRING"],"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/role-memberships":{"get":{"operationId":"RoleMembershipController_getMembershipRoles","parameters":[{"name":"accountType","required":true,"in":"query","schema":{"enum":["EXECUTIVE_ACCOUNT","MEMBER_ACCOUNT"],"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/v1/roles/{code}/memberships":{"post":{"operationId":"RoleMembershipController_addMembership","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleMembershipRequestDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RoleMembershipController_getMember","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/roles/{code}/memberships/{identifier}":{"delete":{"operationId":"RoleMembershipController_removeMember","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}},{"name":"identifier","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/roles":{"post":{"operationId":"RoleController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleRequest"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"RoleController_get","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/roles/{code}":{"delete":{"operationId":"RoleController_delete","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"get":{"operationId":"RoleController_getRole","parameters":[{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/withdrawals/{reference}":{"get":{"operationId":"WalletWithdrawalController_get","parameters":[{"name":"reference","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/wallets":{"get":{"operationId":"WalletController_get","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"WalletController_withdraw","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletWithdrawalDto"}}}},"responses":{"201":{"description":""}}}},"/api/v1/wallets/transactions":{"get":{"operationId":"WalletController_getWalletTransaction","parameters":[{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}},{"name":"minAmountInMinorUnit","required":false,"in":"query","schema":{"type":"number"}},{"name":"maxAmountInMinorUnit","required":false,"in":"query","schema":{"type":"number"}},{"name":"dateCreatedBefore","required":false,"in":"query","schema":{"type":"number"}},{"name":"dateCreatedAfter","required":false,"in":"query","schema":{"type":"number"}},{"name":"type","required":false,"in":"query","schema":{"enum":["CREDIT","DEBIT","WALLET_REVERSAL"],"type":"string"}},{"name":"membershipIdentifier","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/v1/permissions":{"get":{"operationId":"PermissionController_get","parameters":[],"responses":{"200":{"description":""}}}}}}